# backend/Tempswagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trilho API",
    "version": "v1"
  },
  "paths": {
...
    "/api/payments/create-billing": {
      "post": {
        "tags": [
          "PaymentEndpoints"
        ],
        "operationId": "CreateBilling",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/payments/billing/{billingId}": {
      "get": {
        "tags": [
          "PaymentEndpoints"
        ],
        "operationId": "GetBilling",
        "parameters": [
          {
            "name": "billingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/subscription/change-plan": {
      "post": {
        "tags": [
          "PaymentEndpoints"
        ],
        "operationId": "ChangePlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/stations/nearby": {
      "get": {
        "tags": [
          "PingEndpoints"
        ],
        "operationId": "GetNearbyStations",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radiusMeters",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePlanRequest": {
        "type": "object",
        "properties": {
          "planType": {
            "type": "string",
            "nullable": true
          },
          "couponCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateBillingRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "taxId": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "productDescription": {
            "type": "string",
            "nullable": true
          },
          "priceInCents": {
            "type": "integer",
            "format": "int64"
          },
          "isSubscription": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}